requirements_amd.txt
absl-py~=0.10
grpcio~=1.32.0
numpy==1.19.2
paho_mqtt==1.6.1
pandas==1.3.4
protobuf==3.9.2
six~=1.15.0
tensorboard==2.4.0
tensorflow==2.4.0
wheel~=0.35